- package name changed - screen orientation setup moved to manifest file
