+++ app/src/main/AndroidManifest.xml
@@ -1,17 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.chester.puzzlepieces">
+    package="com.example.alexberdnikov.puzzlepieces">
 
   <application
-      android:name=".PuzzleApp"
+      android:name="com.example.alexberdnikov.puzzlepieces.PuzzleApp"
       android:allowBackup="true"
       android:icon="@mipmap/ic_launcher"
       android:label="@string/app_name"
       android:roundIcon="@mipmap/ic_launcher_round"
       android:supportsRtl="true"
       android:theme="@style/AppTheme">
-    <activity android:name=".MainActivity"
-        android:theme="@style/Theme.AppCompat.Light.NoActionBar">
+    <activity android:name="com.example.alexberdnikov.puzzlepieces.MainActivity"
+        android:theme="@style/Theme.AppCompat.Light.NoActionBar"
+        android:screenOrientation="sensorLandscape">
       <intent-filter>
         <action android:name="android.intent.action.MAIN"/>
