--- app/src/main/java/com/example/chester/puzzlepieces/MainActivity.java
@@ -1,28 +1,0 @@
-package com.example.chester.puzzlepieces;
-
-import android.app.ActionBar;
-import android.content.pm.ActivityInfo;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.view.View;
-
-import butterknife.BindView;
-import butterknife.ButterKnife;
-
-public class MainActivity extends AppCompatActivity {
-
-  @BindView(R.id.main_view) PuzzleView puzzleView;
-
-  @Override
-  protected void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_SENSOR_LANDSCAPE);
-
-    View decorView = getWindow().getDecorView();
-    decorView.setSystemUiVisibility(View.SYSTEM_UI_FLAG_FULLSCREEN
-        | View.SYSTEM_UI_FLAG_HIDE_NAVIGATION | View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY);
-
-    setContentView(R.layout.activity_main);
-    ButterKnife.bind(this);
-  }
-}
