+++ app/src/main/res/layout/activity_main.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.example.chester.puzzlepieces.PuzzleView
+<com.example.alexberdnikov.puzzlepieces.PuzzleView
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main_view"
+    android:id="@+id/puzzle_view"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".MainActivity"/>
